did some templating stuff
