+++ app/templates/_base.html
@@ -1,0 +1,18 @@
+<!doctype html>
+<html class="no-js" lang="">
+  <head>
+    {% block head %}
+    <meta charset="utf-8">
+    <meta content="ie=edge" http-equiv="x-ua-compatible">
+    <title>{% block page_title %}eQ - Survey Runner{% endblock %}</title>
+    <meta content="" name="description">
+    <meta content="width=device-width, initial-scale=1" name="viewport">
+    <link href="apple-touch-icon.png" rel="apple-touch-icon">
+    <link href="/static/css/responsive.css" rel="stylesheet">
+    {% endblock %}
+  </head>
+  <body>
+    {% block content %}{% endblock %}
+    <script src="/static/js/bundle.js"></script>
+  </body>
+</html>
