--- app/templates/_base_page.html
@@ -1,22 +1,0 @@
-
-{% block head %}
-{% endblock %}
-
-{% block cookie_message %}
-{% endblock %}
-
-{% block header_class %}{% endblock %}
-
-{% block header %}
-{% endblock %}
-
-{% block content%}
-{% endblock %}
-
-
-
-{% block footer_top %}
-{% endblock %}
-
-{% block body_end %}
-{% endblock %}
