+++ app/templates/errors/404.html
@@ -1,9 +1,9 @@
-{% extends "_base_page.html" %}
+{% extends "_base.html" %}
 
 {% block page_title %}Page could not be found{% endblock %}
 
 {% block content %}
-
-            <h1>Page could not be found</h1>
-    
+  <div class="container">
+    <h1>Page could not be found</h1>
+  </div>
 {% endblock %}
