+++ app/templates/errors/503.html
@@ -1,9 +1,9 @@
-{% extends "_base_page.html" %}
+{% extends "_base.html" %}
 
 {% block page_title %}Page could not be found{% endblock %}
 
 {% block content %}
-
-            <h1>Sorry, we're experiencing technical difficulties</h1>
-
+  <div class="container">
+    <h1>Sorry, we're experiencing technical difficulties</h1>
+  </div>
 {% endblock %}
