+++ app/templates/index.html
@@ -1,25 +1,16 @@
-<!doctype html>
-<html class="no-js" lang="">
-  <head>
-    <meta charset="utf-8">
-    <meta content="ie=edge" http-equiv="x-ua-compatible">
-    <title>eQ - Survey Runner</title>
-    <meta content="" name="description">
-    <meta content="width=device-width, initial-scale=1" name="viewport">
-    <link href="apple-touch-icon.png" rel="apple-touch-icon">
-    <link href="/static/css/responsive.css" rel="stylesheet">
-  </head>
-  <body>
-    <div class="container">
-      <h1>Hello World</h1>
-      <p>This is the eQ Survey Runner (or at least it will be)</p>
-      <p>If you want to know how I'm doing, you can <a href="/healthcheck">check my health</a>.</p>
-      <div class="grid">
-        <div class="grid__col s-4">A</div>
-        <div class="grid__col s-4">B</div>
-        <div class="grid__col s-4">C</div>
-      </div>
+{% extends '_base.html' %}
+
+{% block content %}
+
+  <div class="container">
+    <h1>Hello World</h1>
+    <p>This is the eQ Survey Runner (or at least it will be)</p>
+    <p>If you want to know how I'm doing, you can <a href="/healthcheck">check my health</a>.</p>
+    <div class="grid">
+      <div class="grid__col s-4">A</div>
+      <div class="grid__col s-4">B</div>
+      <div class="grid__col s-4">C</div>
     </div>
-    <script src="/static/js/bundle.js"></script>
-  </body>
-</html>
+  </div>
+
+{% endblock %}
