+++ gulp/bs.js
@@ -1,0 +1,5 @@
+import bs from 'browser-sync'
+
+let browserSync = bs.create()
+
+export default browserSync
